+++ CHANGELOG.md
@@ -6,15 +6,20 @@
 - Filtering issues table
 - Logging and a log viewer in settings
 
-### Completed
+### Added
 
 - Display version in settings, check git tags for updates
+- Summary column in yesterday time tracking widget, widget now spans full dashboard width ([#12](https://github.com/barney-campbell/jira-helper/pull/12))
 
 # 1.1.0
+
+### Added
 
 - Dashboard widget displaying time tracking from the previous working day
 - Kanban-style to do list with ability to link jira issues
 
 # 1.0.0
+
+### Added
 
 - Initial version ported from WPF app